separate community into explore and community actions
