+++ client/src/Tools/_framework/Paths/SiteHeader.tsx
@@ -180,8 +180,8 @@ export function SiteHeader() {
                   <NavLinkTab to="library" dataTest="Library">
                     Library
                   </NavLinkTab>
-                  <NavLinkTab to="community" dataTest="Community">
-                    Community
+                  <NavLinkTab to="explore" dataTest="Explore">
+                    Explore
                   </NavLinkTab>
                   {!user || user.isAnonymous ? (
                     <NavLinkTab to="code" dataTest="Class Code">
@@ -296,8 +296,8 @@ export function SiteHeader() {
                         <NavLinkDropdownTab to="library" dataTest="Library">
                           Library
                         </NavLinkDropdownTab>
-                        <NavLinkDropdownTab to="community" dataTest="Community">
-                          Community
+                        <NavLinkDropdownTab to="explore" dataTest="Explore">
+                          Explore
                         </NavLinkDropdownTab>
                         {!user || user.isAnonymous ? (
                           <NavLinkDropdownTab to="code" dataTest="Class Code">
