+++ client/src/Tools/_framework/Paths/StudentData.tsx
@@ -16,7 +16,7 @@ import {
   Link as ChakraLink,
 } from "@chakra-ui/react";
 import axios from "axios";
-import { DoenetHeading as Heading } from "./Community";
+import { DoenetHeading as Heading } from "../../../Widgets/Heading";
 import { createFullName } from "../../../_utils/names";
 
 type UserData = {
