+++ client/src/index.tsx
@@ -7,10 +7,11 @@ import "@doenet/doenetml-iframe/style.css";
 
 import { MathJaxContext } from "better-react-mathjax";
 import {
-  loader as communityLoader,
-  action as communityAction,
-  Community,
-} from "./Tools/_framework/Paths/Community";
+  loader as exploreLoader,
+  action as exploreAction,
+  Explore,
+} from "./Tools/_framework/Paths/Explore";
+import { action as communityAdminAction } from "./Tools/_framework/Paths/CommunityAdmin";
 import { loader as adminLoader, Admin } from "./Tools/_framework/Paths/Admin";
 import {
   loader as libraryLoader,
@@ -197,32 +198,31 @@ const router = createBrowserRouter([
       {
         path: "/library",
         loader: libraryLoader,
-        // sharing an action with the community page is somewhat intentional
+        // sharing an action with the explore page is somewhat intentional
         // as it shows cards and admins have the same actions that they can perform
-        // on cards as they can on the community page
+        // on cards as they can on the explore page
         // TODO - determine if this is an okay way to share functionality across
         // pages or a bad idea
-        action: communityAction,
+        action: communityAdminAction,
         element: <Library />,
         errorElement: <ErrorPage />,
       },
       {
-        path: "community",
-        loader: communityLoader,
-        action: communityAction,
-        // action: communitySearchAction,
-        element: <Community />,
+        path: "explore",
+        loader: exploreLoader,
+        action: exploreAction,
+        element: <Explore />,
         errorElement: <ErrorPage />,
       },
       {
         path: "admin",
         loader: adminLoader,
-        // sharing an action with the community page is somewhat intentional
+        // sharing an action with the explore page is somewhat intentional
         // as it shows cards and admins have the same actions that they can perform
-        // on cards as they can on the community page
+        // on cards as they can on the explore page
         // TODO - determine if this is an okay way to share functionality across
         // pages or a bad idea
-        action: communityAction,
+        action: communityAdminAction,
         element: <Admin />,
         errorElement: <ErrorPage />,
       },
